Introduced BraveVpnServiceDesktop
Made BraveVpnService common on all platform and added desktop specific
impls to BraveVpnServiceDesktop.
