+++ browser/brave_vpn/brave_vpn_service_factory.h
@@ -9,12 +9,25 @@
 #include "base/memory/singleton.h"
 #include "components/keyed_service/content/browser_context_keyed_service_factory.h"
 
-class BraveVpnService;
 class Profile;
+
+#if defined(OS_WIN) || defined(OS_MAC)
+class BraveVpnServiceDesktop;
+#endif
+
+#if defined(OS_ANDROID)
+class BraveVpnService;
+#endif
 
 class BraveVpnServiceFactory : public BrowserContextKeyedServiceFactory {
  public:
+#if defined(OS_WIN) || defined(OS_MAC)
+  static BraveVpnServiceDesktop* GetForProfile(Profile* profile);
+#endif
+
+#if defined(OS_ANDROID)
   static BraveVpnService* GetForProfile(Profile* profile);
+#endif
   static BraveVpnServiceFactory* GetInstance();
 
   BraveVpnServiceFactory(const BraveVpnServiceFactory&) = delete;
