+++ browser/ui/webui/brave_vpn/vpn_panel_handler.h
@@ -10,7 +10,7 @@
 #include <vector>
 
 #include "brave/components/brave_vpn/brave_vpn.mojom.h"
-#include "brave/components/brave_vpn/brave_vpn_service.h"
+#include "brave/components/brave_vpn/brave_vpn_service_desktop.h"
 #include "mojo/public/cpp/bindings/pending_receiver.h"
 #include "mojo/public/cpp/bindings/receiver.h"
 #include "ui/webui/mojo_bubble_web_ui_controller.h"
@@ -18,6 +18,8 @@
 namespace content {
 class WebUI;
 }  // namespace content
+
+class BraveVpnServiceDesktop;
 
 class VPNPanelHandler : public brave_vpn::mojom::PanelHandler {
  public:
@@ -27,7 +29,7 @@ class VPNPanelHandler : public brave_vpn::mojom::PanelHandler {
   VPNPanelHandler(
       mojo::PendingReceiver<brave_vpn::mojom::PanelHandler> receiver,
       ui::MojoBubbleWebUIController* webui_controller,
-      BraveVpnService* service);
+      BraveVpnServiceDesktop* service);
 
   VPNPanelHandler(const VPNPanelHandler&) = delete;
   VPNPanelHandler& operator=(const VPNPanelHandler&) = delete;
@@ -41,7 +43,7 @@ class VPNPanelHandler : public brave_vpn::mojom::PanelHandler {
  private:
   mojo::Receiver<brave_vpn::mojom::PanelHandler> receiver_;
   ui::MojoBubbleWebUIController* const webui_controller_;
-  BraveVpnService* service_;
+  BraveVpnServiceDesktop* service_;
 };
 
 #endif  // BRAVE_BROWSER_UI_WEBUI_BRAVE_VPN_VPN_PANEL_HANDLER_H_
