+++ components/brave_vpn/brave_vpn_service_desktop.h
@@ -1,0 +1,66 @@
+/* Copyright (c) 2021 The Brave Authors. All rights reserved.
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this file,
+ * You can obtain one at http://mozilla.org/MPL/2.0/. */
+
+#ifndef BRAVE_COMPONENTS_BRAVE_VPN_BRAVE_VPN_SERVICE_DESKTOP_H_
+#define BRAVE_COMPONENTS_BRAVE_VPN_BRAVE_VPN_SERVICE_DESKTOP_H_
+
+#include <string>
+
+#include "base/observer_list.h"
+#include "base/scoped_observation.h"
+#include "brave/components/brave_vpn/brave_vpn_connection_info.h"
+#include "brave/components/brave_vpn/brave_vpn_os_connection_api.h"
+#include "brave/components/brave_vpn/brave_vpn_service.h"
+
+class BraveVpnServiceDesktop
+    : public BraveVpnService,
+      public brave_vpn::BraveVPNOSConnectionAPI::Observer {
+ public:
+  class Observer : public base::CheckedObserver {
+   public:
+    virtual void OnConnectionStateChanged(bool connected) = 0;
+    virtual void OnConnectionCreated() = 0;
+    virtual void OnConnectionRemoved() = 0;
+
+   protected:
+    ~Observer() override = default;
+  };
+
+  explicit BraveVpnServiceDesktop(
+      scoped_refptr<network::SharedURLLoaderFactory> url_loader_factory);
+  ~BraveVpnServiceDesktop() override;
+
+  BraveVpnServiceDesktop(const BraveVpnServiceDesktop&) = delete;
+  BraveVpnServiceDesktop& operator=(const BraveVpnServiceDesktop&) = delete;
+
+  void Connect();
+  void Disconnect();
+  bool IsConnected() const;
+  void CreateVPNConnection();
+  void RemoveVPNConnnection();
+
+  void AddObserver(Observer* observer);
+  void RemoveObserver(Observer* observer);
+
+ private:
+  // BraveVpnService overrides:
+  void Shutdown() override;
+
+  // brave_vpn::BraveVPNOSConnectionAPI::Observer overrides:
+  void OnCreated(const std::string& name) override;
+  void OnRemoved(const std::string& name) override;
+  void OnConnected(const std::string& name) override;
+  void OnDisconnected(const std::string& name) override;
+
+  brave_vpn::BraveVPNConnectionInfo GetConnectionInfo();
+
+  bool is_connected_ = false;
+  base::ObserverList<Observer> observers_;
+  base::ScopedObservation<brave_vpn::BraveVPNOSConnectionAPI,
+                          brave_vpn::BraveVPNOSConnectionAPI::Observer>
+      observed_{this};
+};
+
+#endif  // BRAVE_COMPONENTS_BRAVE_VPN_BRAVE_VPN_SERVICE_DESKTOP_H_
